Update signals.c
The arguments for kill should be in the order (PID, SIGNAL), but it appears to be used as (SIGNAL, PID).
This change ensures that the signal is sent to the process with the ID f_current.pid. Using the correct order of arguments is important for the proper functioning of the kill function.